finished task 2
I used the Vimeo player library.
I added the library as a project dependency via npm.
Used the on() method to track the timeupdate event - updating the playback time.
Save the playback time to the local storage. The key for the storage is the string "videoplayer-current-time".
When the page reloads, I use the setCurrentTime() method to resume playback from the saved position.
I added the lodash.throttle library to the project so that the playback time is updated in the storage no more than once per second.